Custom ParseQueryString
